% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_draft.R
\name{to_numeric}
\alias{to_numeric}
\title{Convert certain variables to numeric values}
\usage{
to_numeric(df, ...)
}
\arguments{
\item{df}{Any data frame you want to edit}

\item{...}{Unquoted variable names (either comma seperated or colon separated for a range of variables)}
}
\value{
A data frame
}
\description{
Converts (by default) 2 to last variables to numeric values. Or use the \code{...} argument to determine which cols to convert.
}
